requirements.txt
Django
djangorestframework
djoser
django-cors-headers
django-filter
